:truck: rename HIntBalloon to HintBallon
