+++ src/components/hint/HintBalloon.tsx
